requirements.txt
fastapi
uvicorn
python-hcl2
python-multipart